requirements.txt
fastapi
uvicorn
line-bot-sdk
python-dotenv
google-generativeai
pdfplumber